Refactored utils to NCUtils
