+++ opensrp-chw-anc/src/main/java/org/smartregister/chw/anc/activity/BaseAncRegisterActivity.java
@@ -39,7 +39,7 @@ import timber.log.Timber;
 
 import static org.smartregister.chw.anc.util.Constants.EVENT_TYPE.PREGNANCY_OUTCOME;
 import static org.smartregister.chw.anc.util.Constants.EVENT_TYPE.UPDATE_EVENT_CONDITION;
-import static org.smartregister.chw.anc.util.Util.saveVaccineEvents;
+import static org.smartregister.chw.anc.util.NCUtils.saveVaccineEvents;
 import static org.smartregister.immunization.ImmunizationLibrary.getInstance;
 import static org.smartregister.util.JsonFormUtils.fields;
 import static org.smartregister.util.JsonFormUtils.getFieldJSONObject;
