+++ opensrp-chw-pnc/src/main/AndroidManifest.xml
@@ -1,2 +1,1 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.smartregister.chw.pnc" />
+<manifest package="org.smartregister.chw.pnc" />
