+++ opensrp-chw-pnc/src/main/java/org/smartregister/chw/pnc/repository/ProfileRepository.java
@@ -46,7 +46,6 @@ public class ProfileRepository extends BaseRepository {
     }
 
 
-
     public List<CommonPersonObjectClient> getChildrenLessThan29DaysOld(String motherBaseEntityID) {
         List<CommonPersonObjectClient> childMemberObjects = new ArrayList<>();
