+++ opensrp-chw-pnc/src/main/java/org/smartregister/chw/pnc/util/Constants.java
@@ -2,13 +2,13 @@ package org.smartregister.chw.pnc.util;
 
 public class Constants {
 
-    public final class KEY {
-
-        public static final String GENDER = "gender";
-    }
-
     public interface TABLES {
         String EC_CHILD = "ec_child";
         String EC_PREGNANCY_OUTCOME = "ec_pregnancy_outcome";
+    }
+
+    public final class KEY {
+
+        public static final String GENDER = "gender";
     }
 }
