+++ opensrp-chw-pnc/src/test/java/org/smartregister/chw/pnc/ExampleUnitTest.java
@@ -2,7 +2,7 @@ package org.smartregister.chw.pnc;
 
 import org.junit.Test;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 /**
  * Example local unit test, which will execute on the development machine (host).
